+++ data/src/main/java/com/youhajun/data/models/dto/store/PurchaseItem.kt
@@ -1,7 +1,7 @@
 package com.youhajun.data.models.dto.store
 
 interface PurchaseItem {
-    val idx: Int?
+    val idx: Long?
     val productId: String?
     val purchaseType: String?
     val price: String?
