+++ domain/src/main/java/com/youhajun/domain/models/vo/PurchaseItemVo.kt
@@ -3,7 +3,7 @@ package com.youhajun.domain.models.vo
 import com.youhajun.domain.models.enums.PurchaseType
 
 interface PurchaseItemVo {
-    val idx: Int
+    val idx: Long
     val productId: String
     val purchaseType: PurchaseType
     val price: String
